requirements.txt
fastapi
uvicorn[standard]
ultralytics
pillow
torch
opencv-python-headless
python-multipart
